+++ app/src/main/java/com/zegoggles/smssync/preferences/Preferences.java
@@ -98,7 +98,6 @@ public class Preferences {
         NOTIFICATIONS("notifications"),
         FIRST_USE("first_use"),
         IMAP_SETTINGS("imap_settings"),
-        DONATE("donate"),
         BACKUP_SETTINGS_SCREEN("com.zegoggles.smssync.activity.fragments.AutoBackupSettings"),
         SMS_DEFAULT_PACKAGE("sms_default_package"),
         SMS_DEFAULT_PACKAGE_CHANGE_SEEN("sms_default_package_change_seen"),